Modif CBean int/double

+++ Flag Catcher/app/src/main/java/com/pokemongo/flagcatcher/model/beans/CoordinateBean.kt
@@ -1,6 +1,6 @@
 package com.pokemongo.flagcatcher.model.beans
 
-data class CoordinateBean ( val id_coordinate: Int,  val long_coordinate: Int, val lat_coordinate: Int){
+data class CoordinateBean ( val id_coordinate: Int,  val long_coordinate: Double, val lat_coordinate: Double){
